local-agent/requirements.txt
fastapi
uvicorn
python-dotenv
openai>=1.0.0
pydantic
google-generativeai
anthropic
mistralai
httpx
pytest
